fix: move task wrappers out of workpool and into workpool/task
