+++ workpool/task/task_test.go
@@ -1,4 +1,4 @@
-package workpool
+package task
 
 import (
 	"context"
